refactor(copyright): extract pattern_extract sub-module from detector.rs
Move regex-based pattern extraction and specialized transform
functions (4216 lines) into detector_pattern_extract.rs. Includes
all extract_*, apply_*, normalize_*, and drop_* functions for
line-pattern matching.
